% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activityFunctions.R
\name{actDataCleaner}
\alias{actDataCleaner}
\title{Cleans up the activity data by converting all datetime values to the
appropriate timezone (currently, probably permanently, Eastern) and then
splitting lines where activities cross days}
\usage{
actDataCleaner(df, tzDF, startVar, endVar, shiftVar, tzVar)
}
\arguments{
\item{df}{The data frame containing my activities data}

\item{tzDF}{The timezone conversion data frame}

\item{startVar}{The column containing the starting datetime for each activity}

\item{endVar}{The column containing the ending datetime for each activity}

\item{shiftVar}{The column in the timezone conversion data frame containing
the number of hours to shift the time by}

\item{tzVar}{The column in the activities data frame containing the text
timezone value.}
}
\value{
Activities data with timezones shifted to the appropriate times
}
\description{
This function converts timezones on the activities data using a separate data
set and then splits lines that cross days into 2 separate line items
}
