% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activityFunctions.R
\name{actDataDateSplitter}
\alias{actDataDateSplitter}
\title{After a timezone conversion, the activities data may have certain activities
that start and end on different days. This is not how the activity data is
intended to be designed - this function splits activities that start and end
on separate days into multiple line items}
\usage{
actDataDateSplitter(df, startVar, endVar, idVar)
}
\arguments{
\item{df}{The data frame containing my activities data}

\item{startVar}{The column containing the starting datetime for each activity}

\item{endVar}{The column containing the ending datetime for each activity}

\item{idVar}{The column containing a unique identifier for each original line
item}
}
\value{
Activities data cross-day line items split into multiple lines
}
\description{
This function splits activities that start and end on different days into 2
separate line items
}
