% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activityFunctions.R
\name{actDataTZCorrection}
\alias{actDataTZCorrection}
\title{This function corrects the timezones in the activities data set using a
separate, outside table. It takes the outside table and adds (or subtracts)
hours from the start and end time of an activity based on the shiftVar, which
should be a numeric value indicating the total hours that the data needs to
move from one timezone to another. Til present, the shift to column has been
eastern time.}
\usage{
actDataTZCorrection(
  df,
  tzDF,
  startVar,
  endVar,
  shiftVar,
  tzVar,
  mergeList = c("timezone", "amDst", "locDst")
)
}
\arguments{
\item{df}{The data frame containing my activities data}

\item{tzDF}{The timezone conversion data frame}

\item{startVar}{The column containing the starting datetime for each activity}

\item{endVar}{The column containing the ending datetime for each activity}

\item{shiftVar}{The column in the timezone conversion data frame containing
the number of hours to shift the time by}

\item{tzVar}{The column in the activities data frame containing the text
timezone value.}

\item{mergeList}{The variables on which to join the timezone shift table to
the activities data. Typically a timezone variable and two indicator
variables that indicate whether or not the secondary timezeon is on
daylight savings}
}
\value{
Activities data with start and end values corrected
}
\description{
This function corrects the start and end variables in the activity data set
to a new timezone.
}
