% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingFunctions.R
\name{correctProjection}
\alias{correctProjection}
\title{Creates a CRS string based on a polygon that corrects the projection of the
polygon}
\usage{
correctProjection(sfDF)
}
\arguments{
\item{sfDF}{An SF object with at least 2 points in it}
}
\value{
A string that can be used to set the CRS of an SF object
}
\description{
This function intakes a polygon and returns a CRS string to best fit that
polygon
}
