% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingFunctions.R
\name{polygonClipper}
\alias{polygonClipper}
\title{This function will take a polygon and cut it off at a specific set of
coordinates. To accomplish this, it uses a dataframe containing columns
titled value (which contains coordinates like xmin, xmax, ymin, and ymax),
and a second column containing the value to cut at.}
\usage{
polygonClipper(polygon, overrideDF)
}
\arguments{
\item{polygon}{a polygon to clip}

\item{overrideDF}{a dataframe containing information on where to clip}
}
\value{
the polygon, clipped at the coordinates
}
\description{
This function cuts a polygon in the specific lat/long value. It does this
using a data frame containing information on where to cut. NOTE: Only works
in lat/long coordinate systems
}
