% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toDoFunctions.R
\name{readToDoRituals}
\alias{readToDoRituals}
\title{Reads in the ritual to do table, which is formatted kind of weirdly, in a
format that is significantly easier to work with}
\usage{
readToDoRituals(
  inFile,
  ritualSheet = "rituals",
  parentSheet = "Ritual parentage",
  deetSheet = "Ritual active range"
)
}
\arguments{
\item{inFile}{The directory to project list.xlsx}

\item{ritualSheet}{The name of the sheet where the ritual table lives}

\item{parentSheet}{The name of the sheet that contains ritual parent data,
which is split out from teh ritual sheet}

\item{deetSheet}{The name of the sheet that contains the ritual details
table, separate from the main sheet}
}
\value{
A cleaned, long-form data set containing the completion times of all
  of my ritual tasks
}
\description{
This function reads the ritual table out of the project list.xlsx file.
}
