% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingFunctions.R
\name{routeCutterLinePoint}
\alias{routeCutterLinePoint}
\title{This function takes a point or line geographic object and then intersects it
with a bbox draw around another geometry, doing everything necessary
including correcting the crs.}
\usage{
routeCutterLinePoint(inputPolygon, cutGeometry, crsMatch, edgeBuffer)
}
\arguments{
\item{inputPolygon}{An overlay polygon that clips the input polygon}

\item{cutGeometry}{The polygon to be intersected with the inputPolygon}

\item{crsMatch}{A Coordinate System that matches both polygons for the cut}

\item{edgeBuffer}{This value expands or contracts the outer, cutting polygon.
Note it will be in whatever metric your crs is in, whether feet, degrees,
or meters.}
}
\value{
the polygon, clipped by the cutting polygon
}
\description{
This polygon will use a polygon to clip point or line geometry
}
