% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingFunctions.R
\name{routeCutterPolygon}
\alias{routeCutterPolygon}
\title{This function takes an input geometry, draws a bounding box geometry around
it, and then then intersects a second geometry with it, doing everything
necessary including correcting the crs and buffering polygons}
\usage{
routeCutterPolygon(
  inputPolygon,
  cutPolygon,
  crsMatch,
  edgeBuffer,
  borderBuffer
)
}
\arguments{
\item{inputPolygon}{An overlay polygon that clips the input polygon}

\item{cutPolygon}{The polygon to be intersected with the inputPolygon}

\item{crsMatch}{A Coordinate System that matches both polygons for the cut}

\item{edgeBuffer}{This value expands or contracts the outer, cutting polygon.
Note it will be in whatever metric your crs is in, whether feet, degrees,
or meters.}

\item{borderBuffer}{This value expands or contracts the final polygon.}
}
\value{
the polygon, clipped by the cutting polygon
}
\description{
This polygon will use one polygon to clip another.
}
