% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingFunctions.R
\name{subwayRoutes}
\alias{subwayRoutes}
\title{Uses 2 input data frames to create specific subway and train travel lines
that I have traversed in a given timeframe. One data frame contains dates and
stops, another contains subway route data, broken up by traversals between
two stops}
\usage{
subwayRoutes(startStopDF, subwaySFDF)
}
\arguments{
\item{startStopDF}{a dataframe containing the starting trainstop and ending
trainstop for a given train ride}

\item{subwaySFDF}{a geographic dataframe containing all of the subway lines
in a given geographic area, where the sf data is broken down based on legs
between stops}
}
\value{
A dataframe containing all of the train routes between two stops
}
\description{
This function intakes two data frames and builds out a geographic dataframe
containing all of the routes that were developed between two routes.
}
