% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activityFunctions.R
\name{tzCorrecter}
\alias{tzCorrecter}
\title{A function to shift the timezone of a variable based on an outside table. It
takes the outside table and adds (or subtracts) hours from chosen variable
based on the shiftVar, which should be a numeric value indicating the total
hours that the data needs to move from one timezone to another. Til present,
the shift to column has been eastern time.}
\usage{
tzCorrecter(df, tzDF, tzVar, timeVar, shiftVar, mergeList)
}
\arguments{
\item{df}{The data frame containing my activities data}

\item{tzDF}{The timezone conversion data frame}

\item{tzVar}{The column in the activities data frame containing the text
timezone value.}

\item{timeVar}{The column containing datetime to shift}

\item{shiftVar}{The column in the timezone conversion data frame containing
the number of hours to shift the time by}

\item{mergeList}{The variables on which to join the timezone shift table to
the activities data. Typically a timezone variable and two indicator
variables that indicate whether or not the secondary timezone is on
daylight savings}
}
\value{
Activities data with start and end values corrected
}
\description{
Shifts a single variable to a new timezone based on a separate data table
}
